refactor: change file names to stay consistent with tech specs
