+++ first-project/src/__tests__/card.test.tsx
@@ -1,5 +1,5 @@
 import { render, screen } from '@/__tests__/test-utils/test-utils';
-import CardItem from '@/components/cards-list/cards-item';
+import CardItem from '@/components/cards-list/card';
 import { mockCardData, incompleteCardData } from '@/__tests__/mocks/handlers';
 
 test('CardItem displays item name and description correctly', () => {
