+++ first-project/src/components/cards-list/card.tsx
